Hibernate was swap with JPA. Small issues were solved.

+++ src/main/java/ua/lviv/lgs/controllers/admin/AdminProductsController.java
@@ -2,7 +2,9 @@ package ua.lviv.lgs.controllers.admin;
 
 import org.apache.log4j.Logger;
 import ua.lviv.lgs.domain.Product;
+import ua.lviv.lgs.service.BucketService;
 import ua.lviv.lgs.service.ProductService;
+import ua.lviv.lgs.service.impl.BucketServiceImpl;
 import ua.lviv.lgs.service.impl.ProductServiceImpl;
 
 import javax.servlet.ServletException;
@@ -18,9 +20,11 @@ public class AdminProductsController extends HttpServlet {
 
     private static final Logger log = Logger.getLogger(ua.lviv.lgs.controllers.ProductsController.class);
 
+    private final BucketService bucketService;
     private final ProductService productService;
 
     public AdminProductsController() {
+        this.bucketService = BucketServiceImpl.getInstance();
         this.productService = ProductServiceImpl.getInstance();
     }
 
@@ -38,6 +42,12 @@ public class AdminProductsController extends HttpServlet {
         log.info("DELETE: remove product from list");
 
         int productId = Integer.parseInt(req.getPathInfo().substring(1));
+
+        if (bucketService.containsProductById(productId)) {
+            resp.setStatus(403);
+            return;
+        }
+
         productService.delete(productId);
     }
 }
